replicate_deployment/deepseekocr/requirements.txt
transformers>=4.30.0
torch>=2.0.0
pillow>=9.0.0
cog>=0.9.0
